Change the property to match the width of .footer
Flex: 1 will adjust itself based on the
display: flex property on the container.
This result that the items will stretch
to match the width of the footer which stretch
the items equally

Flex: 1 was used to display the items because
width is not defined, therefore if flex: 1 isnt
declared and display: flex is enabled in the
container, it will show no items.